Modify gameday text logic (#27)
* Modify gameday text logic

Show short month plus numeric day if game not occurring within a week.
Show "Today" if the game is occurring today. This helps avoid any
confusion at first glance when the day of the week is shown.

Note: had trouble with the translate function of the Translator, so
hardcoded this English. I'm not familiar with this, so may be an easy
fix to have this display properly and include locale support.

* gameday display: Adjust style for short date

Added div type for gamedateshort and matched it to the smaller font size
of the game time.

* Rewrite gameday display logic

Clean up implementation to ensure corner cases are covered and properly
implement translation.